+++ cache-config/t3c-apply/t3c-apply.go
@@ -59,6 +59,10 @@ func runSysctl(cfg config.Cfg) {
 	}
 }
 
+const LockFilePath = "/var/run/t3c.lock"
+const LockFileRetryInterval = time.Second
+const LockFileRetryTimeout = time.Minute
+
 func main() {
 	var syncdsUpdate torequest.UpdateStatus
 	var lock util.FileLock
@@ -69,9 +73,16 @@ func main() {
 	} else if cfg == (config.Cfg{}) { // user used the --help option
 		os.Exit(Success)
 	}
-	if !lock.GetLock("/var/run/t3c.lock") {
-		os.Exit(AlreadyRunning)
+
+	log.Infoln("Trying to acquire app lock")
+	for lockStart := time.Now(); !lock.GetLock(LockFilePath); {
+		if time.Since(lockStart) > LockFileRetryTimeout {
+			log.Errorf("Failed to get app lock after %v seconds, another instance is running, exiting without running\n", int(LockFileRetryTimeout/time.Second))
+			os.Exit(AlreadyRunning)
+		}
+		time.Sleep(LockFileRetryInterval)
 	}
+	log.Infoln("Acquired app lock")
 
 	if cfg.UseGit == config.UseGitYes {
 		err := util.EnsureConfigDirIsGitRepo(cfg)